refactor: import supportsFiles and models vars from schemas
