+++ api/server/routes/endpoints/schemas.js
@@ -11,6 +11,30 @@ const EModelEndpoint = {
   assistant: 'assistant',
 };
 
+const supportsFiles = {
+  [EModelEndpoint.openAI]: true,
+  [EModelEndpoint.assistant]: true,
+};
+
+const openAIModels = [
+  'gpt-3.5-turbo-16k-0613',
+  'gpt-3.5-turbo-16k',
+  'gpt-4-1106-preview',
+  'gpt-3.5-turbo',
+  'gpt-3.5-turbo-1106',
+  'gpt-4-vision-preview',
+  'gpt-4',
+  'gpt-3.5-turbo-instruct-0914',
+  'gpt-3.5-turbo-0613',
+  'gpt-3.5-turbo-0301',
+  'gpt-3.5-turbo-instruct',
+  'gpt-4-0613',
+  'text-davinci-003',
+  'gpt-4-0314',
+];
+
+const visionModels = ['gpt-4-vision', 'llava-13b'];
+
 const eModelEndpointSchema = z.nativeEnum(EModelEndpoint);
 
 const tPluginAuthConfigSchema = z.object({
@@ -353,4 +377,7 @@ module.exports = {
   parseConvo,
   getResponseSender,
   EModelEndpoint,
+  supportsFiles,
+  openAIModels,
+  visionModels,
 };
